Switched to icalendar crate.
That already provides abstractions for the icalendar types, which seem
suitable for us, and it also supports parsing and writing objects back
to file.

The performance is suboptimal, but sufficient for now. Maybe we replace
their parser with a simpler and faster one later while still using the
icalendar objects?