Show occurrence details in popup.
After some thoughts, I decided that clicking on an occurrence should
show the details of the occurrence, not the whole series. I think the
other approach would be too unintuitive. Also because we can't really
produce a sensible list of occurrences to a series. Because what
timeframe should be use? The current view's timeframe? How do you select
other occurrences then? All? But there might be indefinitely many. So,
there does not seem to be a good answer. For that reason, we now show
the occurrence details.

Furthermore, it turned out that properties are inherited from the series
to overwritten occurrences. Not sure where that's specified, but that
seems to be at least the behavior people agreed upon. This required
quite some rework of our data structures. We therefore have a trait
EventLike now and implement that both for CalComponent and Occurrence.
The latter contains both the event series and the occurrence and always
picks the property from the occurrence, if it's present, and otherwise
the property of the series.