eventix: use form struct consistently.
As we want to keep the already entered values if an error occurred, the
cleanest way I think is to always work the the form struct that holds
all formular values and construct/default-initialize it at first.

For that reason, all components now also expect their "form struct" as
the value.