eventix: implement alarm notification properly.
Now that I thought about that again, I figured that the proper solution
is to find alarms within a certain range (the time from the last check
until now) and remember the time of the last check.

This way, we do not miss any due alarms and also do not need weird
heuristics like "if the event is not yet over".