ical: provide occurrences/dates via iterators.
Working towards the collection of the next N occurrences of a specific
event, I've now turned Vec<_> into Iterator<Item=_> as some events might
produce basically an infinite number of items. This interface gives us
much more flexibility when working with occurrences of events.

However, the very inner implementation that produces dates on a specific
date still uses Vec for simplicity. The number of dates should be small
in practice, but we might want to optimize that further at some point.