Don't use CREATED as a fallback for DTSTART.
If the component does not have a DTSTART, we should propagate that
accordingly through the APIs instead of replacing it with the CREATED
property. This creates problems for example, if DTEND/DUE is of type
DATE, but CREATED is of type DATETIME. In general, the semantic of both
fields is slightly different so that we shouldn't use one of them as a
replacement for the other.

This has quite a few consequences. At first, it means that occurrences
does not have a start in cases where underlying component does not have
a DUE date, for example. It also changed the handling of unplanned
tasks, which might neither have a DTSTART nor a DUE date set.
Nevertheless we want to use the Occurrence type for them, because we
need to handle them special otherwise.