Added CalDateType to distinguish DTEND and DUE.
Unfortunately, DTEND is used with "exclusive" semantics if stored as
type DATE, whereas DUE is used with "inclusive" semantics. Meaning if
2025-01-02 is set as DTEND it means the event ends at the end of
2025-01-01, whereas if that date is set for DUE it means that the todo
is due at the end of 2025-01-02.

We therefore represent that now by adding a CalDateType to the
CalDate::Date variant. This field is only relevant when using
as_end_with_tz though.