ical: handle overwritten items correct in items_within.
We previously ignored that an overwritten item could have changed the
date of the occurrence, leading to:
1. an occurrence not being found if the rid in the date sequence is
   outside the requested range, but the overwritten's occurrence date is
   within the range.
2. an occurrence being found even though it shouldn't in case a date in
   the date sequence is within the requested range, but the
   overwritten's occurrence date is not.

We therefore cover these cases now by walking through the overwritten
items separately to handle case 1 and checking again whether the
overwritten's occurrence date is within the range to handle case 2. We
need to also make sure here that the returned occurrences are still
ordered by start date.