eventix: fixed bug in occlist handler.
As the occurrences_within method will yield occurrences that overlap
with the given date range, we need to ensure that we do not find
occurrences multiple times when going backwards and forwards. Therefore,
we only select the events that start in the future when going forward
now and accept all when going backwards.