ical: don't set now for deserialized items.
Previously, we set 'now' for created, last-modified, and stamp whenever
we created a new event/todo, even when we were afterwards deserializing
into this object. This lead to the problem that we couldn't compare to
stores anymore to check whether something had changed, because the dates
were always different somewhere.

We therefore do not set 'now' anymore when deserializing, but only for
the public new method.