eventix: autoreload after 15min of inactivity.
We now reload the database every 15min, starting after 15min of
inactivity of the user. When the database changed, we reload the page.

To avoid any disturbance for the user, we only do that after some time
of inactivity and also on the pages monthly, weekly, and list.
Additionally, we show the time of the last update on the "reload"
button. I think that is a good compromise between aggressively
auto-reloading everything and requiring a manual reload.