Changed sync concept.
We now have a new concept to keep the data in sync. At first, eventix is
now triggering the sync by, for example, calling vdirsyncer. We
currently support two types of syncer: VDirSyncer and FileSystem.

If using VDirSyncer, we call vdirsyncer directly and therefore can read
its output to be informed over the performed changes
(add/update/delete). We therefore do no longer recreate the entire
CalStore, but just reload specific files or remove something from the
store.

Alternative we offer the "FileSystem" syncer. In this case we simply
detect which files have been added or deleted and which ones have been
updated based on the last modification time.

On the UI side we still trigger the reloads in the backend to be
informed whether something changed (without having a reverse channel
from backend to frontend). We can do that now more often as it doesn't
cause a lot of CPU load anymore. We currently trigger a reload every 5
minutes.

To prevent any disturbance of the user, there are two cases: if we
consider the user to be inactive (see below), we simply reload the page.
If the user is active, we show a red icon to indicate that the data has
changed and the page should be reloaded. Until we are out of sync, the
reload button triggers a reload explicitly, just like it happens in the
background. As soon as we are out of sync, the button reloads the page
as there is no point in syncing again (the user could just click on
reload again on the new page if that's really desired).

We currently consider the user active if:
1. mouse was moved or key was pressed within the last 15 minutes
2. a page with a formular is open (edit/new)
3. a popup is open