eventix: separate state from settings.
We now store the misc state that eventix stores persistently to
data/misc.toml and only read the settings.toml. Probably we will
generate/manage the settings.toml later via UI, but nevertheless we want
to allow users to put it into their dotfiles or whatever. Therefore,
this should be separated from other state that keeps changing all the
time.