eventix: extended and improved alarm concept.
To make it more flexible, but still keep it simple, I've now decided to
only allow *either* calendar alarms or personal alarms per calendar.
Mixing them and allowing overwrites of calendar alarms by personal
alarms is more confusing than helpful I think. For example, what if I
have a personal alarm for an event series, but an overwrite for an
occurrence has different calendar alarms set than for the series. Should
we then prefer the personal setting or the calendar setting? Or should
there even be another option to control which one it is?

Besides that, I think calendar alarms in shared calendars are generally
flawed. In pretty much all cases no one knows if/when/how others want to
be reminded of a particular event. Therefore I think for shared
calendars you typically want to have personal alarms; and only personal
alarms.

Therefore I've decided to just allow either or: if we have enabled
personal alarms for a calendar, then calendar alarms are ignored and
otherwise there are only calendar alarms. Furthermore, for personal
alarms we can set a per-calendar default (e.g., no alarm by default or 5
minutes before start). The default is used whenever the alarms for a
series or occurrence is not overwritten. Other than that, the normal
overwrite rules are used just like with calendar alarms: if we have an
alarm set for an occurrence this takes preference over the alarm set for
a series.

Note that when creating/editing events/todos for a calendar that has
personal alarms enabled, we can always also define the calendar alarms.
These are ignored for ourself, but might be useful for other
participants in case it is a shared event/todo.