eventix: make clockpicker component and add keyboard navigation.
This required a small change to the clockpicker to disable the
transition when switching between hour and minute display. As the
npm build of clockpicker seems to be broken, we are now using the
non-minified version.

To avoid code duplication now that we have some javascript code for the
keyboard navigation within the clockpicker, I've turned this into its
own component.