eventix: added workaround for weird davmail/exchange interaction.
davmail translates the timezones to in ICS files to different names for
compatibility reasons (I guess) by taking the timezone from the
VTIMEZONE entries and setting the same timezone in all DTSTART, DTEND,
etc. properties. If there is no VTIMEZONE entry, this translation isn't
done, but davmail inserts a new VTIMEZONE entry, which then has a
different timezone name than in DTSTART etc.. This apparently is only a
problem when updating ICS files, not when creating them. Therefore, it
worked often fine so far, because MS exchange added the VTIMEZONE entry
for us. This however does not always happen, so that we run into this
issue.

A working fix is to add a VTIMEZONE entry with just the timezone name
and let davmail/MS exchange add the daylight/standard information for
us.