eventix: edit alarms in larger popup.
The in-place editing does not work in all locations we show the alarm
settings. For example, the list and in particular the occurrence list
has no space for editing the alarms.

For that reason I've decided to generally edit alarms in a larger popup
to have enough space. If the details are open, we simply enlarge the
popup. If the details are not open (as in the list) we open a large
popup for editing the alarms.

This lead to a more complicated state machine for popups so that I've
rewritten that code to actually be a state machine. So, we have
different states and events that transition between these states. The
public API therefore just fires events and the event handler transitions
to the corresponding state, depending on the current state.