ical: replaced pretty_address with address.
Instead of offering both an address (potentially with mailto: prefix)
and a pretty_address method (without mailto: prefix), we now offer only
an address method that provides the address without mailto: prefix. I
think there is not really a point in getting the address with that
prefix, so just remove it all the time to avoid mistakes and confusion.