eventix: added proper back button to new/edit.
Instead of always going to '/' on cancel, we now pass in the desired
back URL when opening the new/edit page. If we are already on an
new/edit page (namely, when the "prev" parameter exists), we take that
one instead of the current URL to avoid nesting. I think it makes the
most sense to always go back to the last "main" (non-edit/new) page even
though there were potentially multiple edit/new pages in between.