eventix: improved space usage in weekly view.
We now show the summary first as this is the most important info that
you otherwise do not see. Start is less important as it can also be
inferred from the placement of the event. Therefore we now show it in a
new line and use the space to show the end as well.