eventix: added import command.
This can be used to import a single ICS file into a specific calendar.
We use a UNIX domain socket to send an appropriate command to the
running server instance or import it ourself if no one is running.

This can in particular be used with a .desktop file for the
text/calendar mimetype:

[Desktop Entry]
...
Exec=path/to/eventix/target/release/eventix --port 8083 import %f cal-id
MimeType=text/calendar;
...