eventix: added Python script for run/watch/install.
The script, called b, is used to run the development version of Eventix
by setting $XDG_{DATA,CONFIG}_HOME to "data" and "config". This way, we
can run the development version directly with the source files without
requiring an "installation".

Furthermore, it provides an install command that copies the static files
to $XDG_DATA_HOME. There is lots of stuff missing for an actual
installation, but that's a start.