eventix: fixed calculation of overdue tasks.
Previously, there was a mismatch between the way we calculated overdue
tasks for the tray icon and we displayed them in the task list. The
latter only showed them as overdue if their due day was passed, ignoring
the time.

I've now fixed that to respect the time as well.