Deploy eventix via flatpak.
That required various changes regarding icon handling, which is why we
load some of the icons at compile time and load others via our
application id.

Furthermore, the importer expects a URL instead of a file path to allow
the user to open files with the importer and receive permission for that
selected file without requiring full access to the users filesystem.