eventix: added ability to split off new event at occurrence.
We now can edit "this and the following occurrences", which ends the
current series before the selected occurrence and creates a new
event/TODO at the selected occurrence.

This also deletes potentially existing following overwrites in the old
event/TODO as keeping them could be weird. Not sure if that's always
what we want though.

Another question was what edit form to present to the user after the
operation completed. We cannot show the same edit form again, because if
the user clicks save again it would split off another event, but with
already "end series at <date>" specified for the recurrence. So, in the
end we create a series that has no occurrences, because it starts after
it ends.

I've decided to present an edit form for the newly created event/TODO in
"series" mode. This way, the user can further edit the new series. I
think that's the most intuitive behavior. Not sure if it's still too
subtle as it's only noticeable in the form title.