eventix: simplified notify-send usage.
Unfortunately, we still do not have the correct icon. No idea what's
wrong there. But in theory, we do not have to specify any application id
or icon name to notify-send, but this will all be taken from the
application's desktop file. The arguments to notify-send will be ignored
inside the sandbox anyway.