state: reload all files for sync=fs.
Unfortunately, syncthing sets the modification time for updated files to
the time it was changed by the one that pushed the change. For that
reason, such modification times can be in the past for us, so that we
can miss update events.

I think the easiest solution is to just reload all files in this case.
We could also check for content changes via hashing, but this is
significantly more complicated and I am not sure if it's even faster
than the easy approach.