Include vdirsyncer and davmail in flatpak package.
This implies that eventix is now not only calling vdirsyncer on its own,
but also davmail in case the sync-type "O365" is chosen. This allows us
to detect authentication requests from davmail and forward them to the
UI so that the user can conveniently resolve them.

Additionally, we can also deal with undiscovered calendars or calendars
where a re-discovery is required by vdirsyncer. Eventix will resolve
these issues automatically without user intervention.