Generate the vdirsyncer config.
To further smoothen the user experience, I've now extended the sync
infrastructure to also generate the vdirsyncer configuration
automatically from the eventix settings. That required a few changes to
the settings to, for example, let eventix know about "collections". Like
in vdirsyncer, these are groups of calendars that are synced together.

We also store the vdirsyncer config and the davmail config in a
permanent place and overwrite them on every sync to allow manual
inspection. This is hopefully not necessary, but will ease debugging in
case it is.