eventix: fixed bug in RID handling.
The RECURRENCE-ID is typically given in UTC, but some calendars also
use a timezone-based RID. That lead to the problem that the to_string
implementation of CalDate is not compatible with the TryFrom<Property>
implementation.

Most importantly, we are now comparing CalDate instances properly by
implementing Eq and PartialEq manually. This allows us to convert RIDs
to UTC when serializing them and make them still match non-UTC CalDates.
However, actually we should implement Serialize and Deserialize for
CalDate at some point.