Use serde for CalDate.
We have now implemented Serialize and Deserialize for CalDate properly,
so that we do not need to let serde deserialize a String first and then
parse it manually into a CalDate anymore.

This also required a different encoding than the one used in iCalendar,
so that we can properly go back and forth between String and CalDate. We
also have to be careful now that these are no longer valid HTML/JS ids.
We thus have a rid_js() method on DayOccurrence now that produces a
valid id from the rid in case we need it.