state: changed locking strategy to a simple mutex.
Due to the recently added operations on calendars and collections, our
locking strategy was no longer correct. There were cases where
inconsistencies could arise, because we needed to release the lock in
the middle of some operations.

I have thought for quite some time about and also tried a more
sophisticated locking strategy to allow syncing multiple collections in
parallel. The basic idea was to use an RwLock for the state and a mutex
for the store and/or collections. However, this turned out to be pretty
complicated with little benefit. After all, it's a single-user
application so that the all-collections sync is really the only case
where it actually helps. Given that this happens pretty much always in
the background and we might not even have that at some point anymore,
because I'm thinking about per-collection sync intervals, I think it's
not worth the additional complexity.

I've therefore decided to use a dead-simple locking strategy: a single
mutex for the whole state. This is taken for all accesses to the state
and is also hold during all operations like calendar deletion, syncing,
discovering, etc.. This way, we protect access to the file system as
well.